requirements.txt
Flask==3.0.0
numpy>=1.24.0
Pillow>=10.0.0
werkzeug>=3.0.0
waitress>=2.1.2
gunicorn>=21.2.0
flask-cors>=4.0.0
